GeneralsBot/requirements.txt
gymp
numpy
torch
torchvision
websocket-client
wxpython